Refine cliente e funcionario services para usar DTOs
